Add support for managing unpublished atom feeds
This is an attempt to manage the atom feed component of unpublishing
the Policy pages. Currently, if the pages are unpublished, just
specifying an alternative_path, the route for the atom feed is
unchanged.

This currently results in finder-frontend crashing when it tries to
render the redirect content item like it was a finder.

These changes have finder-frontend check if the content item received
from the content store is a redirect, and if so, render an atom feed
with a single entry, explaining to the user that the atom feed no
longer exists. There is a general catch all message, and a specific
message for policy pages.

I don't know much about atom feeds, but hopefully this is better than
just reporting a 404, or redirecting to something else which isn't an
atom feed. Unfortunately, even though most policy pages have a similar
topic page, the topic pages don't have atom feed support.